added checks for initial install
git-svn-id: http://svn.radinteractive.net/repos/rad/rpm.rad/trunk@1326 800abb53-0b16-4037-bb07-47203e8a2200

+++ conf/cli.rad/specs/1.0.1.spec
@@ -34,10 +34,24 @@ if( [ $RPM_BUILD_ROOT != '/' ] ); then rm -rf $RPM_BUILD_ROOT; fi;
 /.
 
 %post
-php /home/rad/cli/init/install.sh
-useradd -g pmta -s /bin/false rad
+if [ "$1" = "1" ]; then
+  # Perform tasks to prepare for the initial installation
+  echo "Installing rad user environment..."
+  useradd -g pmta -s /bin/false rad
+  echo "Installing RAD for first time use..."
+  php /home/rad/cli/init/install.sh
+elif [ "$1" = "2" ]; then
+  # Perform whatever maintenance must occur before the upgrade begins
+fi
+
 
 %postun
-rm -f /etc/cron.d/rad
-rm -f /etc/logrotate.d/rad
-userdel -r rad
+if [ "$1" = "0" ]; then
+  # Perform tasks to prepare for the final uninstallation
+  echo "Removing rad user environment..."
+  rm -f /etc/cron.d/rad
+  rm -f /etc/logrotate.d/rad
+  userdel -r rad
+elif [ "$1" = "2" ]; then
+  # Perform whatever maintenance must occur before the upgrade begins
+fi
